+++ src/Search.php
@@ -283,7 +283,7 @@ class Search
      * @throws ServerResponseException  if the status code of response is 5xx
      * @throws ClientException
      */
-    public function search($query = '', ?array $options = null, string $method = Request::POST): ResultSet
+    public function search($query = '', ?array $options = null): ResultSet
     {
         $this->setOptionsAndQuery($options, $query);
 
@@ -323,7 +323,7 @@ class Search
      *
      * @phpstan-return ($fullResult is false ? int : ResultSet)
      */
-    public function count($query = '', bool $fullResult = false, string $method = Request::POST)
+    public function count($query = '', bool $fullResult = false)
     {
         $this->setOptionsAndQuery(null, $query);