+++ src/SearchableInterface.php
@@ -43,7 +43,6 @@ interface SearchableInterface
      * @phpstan-param TCreateQueryArgs $query
      *
      * @param array<string, mixed>|null $options associative array of options (option=>value)
-     * @param string                    $method  Request method, see Request's constants
      *
      * @throws NoNodeAvailableException if all the hosts are offline
      * @throws ClientResponseException  if the status code of response is 4xx
@@ -51,7 +50,7 @@ interface SearchableInterface
      * @throws InvalidException
      * @throws ClientException
      */
-    public function search($query = '', ?array $options = null, string $method = Request::POST): ResultSet;
+    public function search($query = '', ?array $options = null): ResultSet;
 
     /**
      * Counts results for a query.
@@ -62,8 +61,6 @@ interface SearchableInterface
      *
      * @phpstan-param TCreateQueryArgsMatching $query
      *
-     * @param string $method Request method, see Request's constants
-     *
      * @throws NoNodeAvailableException if all the hosts are offline
      * @throws ClientResponseException  if the status code of response is 4xx
      * @throws ServerResponseException  if the status code of response is 5xx
@@ -71,7 +68,7 @@ interface SearchableInterface
      *
      * @return int number of documents matching the query
      */
-    public function count($query = '', string $method = Request::POST);
+    public function count($query = '');
 
     /**
      * @param AbstractQuery|AbstractSuggest|array|Collapse|Query|string|Suggest|null $query